requirements.txt
# Core LLM and Agent Framework
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
openai==1.6.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.2

# Vector Database and Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Data Validation and Models
pydantic==2.5.3
pydantic-settings==2.1.0

# API Framework (Optional)
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Report Generation
